requirements.txt
APScheduler==3.7.0
astroid==2.7.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
cycler==0.10.0
greenlet==1.1.0
grpcio==1.38.1
grpcio-tools==1.38.1
idna==3.2
isort==5.9.3
Jinja2==3.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
numpy==1.21.1
Pillow==8.3.1
platformdirs==2.2.0
protobuf==3.17.3
pylint==2.10.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
readerwriterlock==1.0.8
requests==2.26.0
retrying==1.3.3
six==1.16.0
some-package==0.1
SQLAlchemy==1.4.22
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
win32-setctime==1.0.3
wrapt==1.12.1
